Migrate audio management to footer component
