+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "webplayer",
-  "version": "0.2.1",
+  "version": "1.1.0",
   "license": "GPL-3.0 License",
   "email": "contact@jelliti.me",
   "url": "https://oussama.jelliti.me",
